+++ tests/StyleTest.php
@@ -18,12 +18,7 @@ class StyleTest extends TestCase
      */
     protected $style;
 
-    /**
-     * @var StreamOutput
-     */
-    private $output;
-
-    public function setUp()
+    public function setUp(): void
     {
         $input = new ArgvInput;
         $this->output = new StreamOutput(fopen('php://memory', 'w+', false));
@@ -31,25 +26,25 @@ class StyleTest extends TestCase
         $this->style = new Style($input, $this->output, $formatter);
     }
 
-    public function testWarningMessage()
+    public function testWarningMessage(): void
     {
         $this->style->warningMessage('Warning message');
         $this->assertEquals($this->getDisplay(), "[WARNING]            Warning message\n");
     }
 
-    public function testErrorMessage()
+    public function testErrorMessage(): void
     {
         $this->style->errorMessage('Error message');
         $this->assertEquals($this->getDisplay(), "[ERROR]              Error message\n");
     }
 
-    public function testOkMessage()
+    public function testOkMessage(): void
     {
         $this->style->okMessage('Ok message');
         $this->assertEquals($this->getDisplay(), "[OK]                 Ok message\n");
     }
 
-    public function testInfoMessage()
+    public function testInfoMessage(): void
     {
         $this->style->infoMessage('Info message');
         $this->assertEquals($this->getDisplay(), "[INFO]               Info message\n");